01_blockchain-backend/requirements.txt
Flask==2.2.2
pycryptodome==3.15.0
